Move additional fields
