+++ datafusion/proto-common/proto/datafusion_common.proto
@@ -487,3 +487,49 @@ message CsvWriterOptions {
   // Optional value to represent null
   string null_value = 8;
 }
+
+message Precision{
+  PrecisionInfo precision_info = 1;
+  datafusion_common.ScalarValue val = 2;
+}
+
+enum PrecisionInfo {
+  EXACT = 0;
+  INEXACT = 1;
+  ABSENT = 2;
+}
+
+message Statistics {
+  Precision num_rows = 1;
+  Precision total_byte_size = 2;
+  repeated ColumnStats column_stats = 3;
+}
+
+
+message ColumnStats {
+  Precision min_value = 1;
+  Precision max_value = 2;
+  Precision null_count = 3;
+  Precision distinct_count = 4;
+}
+
+enum JoinSide{
+  LEFT_SIDE = 0;
+  RIGHT_SIDE = 1;
+}
+
+enum JoinType {
+  INNER = 0;
+  LEFT = 1;
+  RIGHT = 2;
+  FULL = 3;
+  LEFTSEMI = 4;
+  LEFTANTI = 5;
+  RIGHTSEMI = 6;
+  RIGHTANTI = 7;
+}
+
+enum JoinConstraint {
+  ON = 0;
+  USING = 1;
+}
