+++ datafusion/proto-common/src/lib.rs
@@ -56,6 +56,7 @@ pub mod to_proto;
 pub use from_proto::Error as FromProtoError;
 pub use generated::datafusion_proto_common as protobuf_common;
 pub use generated::datafusion_proto_common::EmptyMessage;
+pub use generated::datafusion_proto_common::JoinSide;
 pub use to_proto::Error as ToProtoError;
 
 #[cfg(doctest)]
